+++ Driver.h
@@ -1,0 +1,47 @@
+#include <iostream>
+using namespace std;
+namespace TelCoColorCoder
+{
+	enum MajorColor { WHITE, RED, BLACK, YELLOW, VIOLET };
+	enum MinorColor { BLUE, ORANGE, GREEN, BROWN, SLATE };
+	const char* MajorColorNames[] = { "White", "Red", "Black", "Yellow", "Violet" };
+	int numberOfMajorColors =sizeof(MajorColorNames) / sizeof(MajorColorNames[0]);
+	const char* MinorColorNames[] = {"Blue", "Orange", "Green", "Brown", "Slate"};
+	int numberOfMinorColors =sizeof(MinorColorNames) / sizeof(MinorColorNames[0]);
+
+	class ColorPair {
+	private:
+		MajorColor majorColor;
+		MinorColor minorColor;
+	public:
+		ColorPair(MajorColor major, MinorColor minor) :
+			majorColor(major), minorColor(minor)
+		{}
+		MajorColor getMajor() {
+			return majorColor;
+		}
+		MinorColor getMinor() {
+			return minorColor;
+		}
+		
+        string ToString() 
+        {
+		string colorPairStr = MajorColorNames[majorColor];
+		colorPairStr += " ";
+		colorPairStr += MinorColorNames[minorColor];
+		return colorPairStr;
+		}
+	};
+
+	ColorPair GetColorFromPairNumber(int pairNumber) {
+		int zeroBasedPairNumber = pairNumber - 1;
+		MajorColor majorColor =
+			(MajorColor)(zeroBasedPairNumber / numberOfMinorColors);
+		MinorColor minorColor =
+			(MinorColor)(zeroBasedPairNumber % numberOfMinorColors);
+		return ColorPair(majorColor, minorColor);
+	}
+	int GetPairNumberFromColor(MajorColor major, MinorColor minor) {
+		return major * numberOfMinorColors + minor + 1;
+	}
+}
